fix: show patient admit button on reservation rows
Normalize booking status so admit/no-show actions appear; remove Today button; open prefilled admission modal.

+++ Backend/Healan/src/Healan.Application/Booking/Queries/PortalBooking/PortalBookingQueries.cs
@@ -216,7 +216,7 @@ public class PortalMyBookingsQueryHandler : IRequestHandler<PortalMyBookingsQuer
                 FirstName = x.FirstName,
                 LastName = x.LastName,
                 Note = x.Note,
-                Status = x.Status,
+                Status = (byte)x.Status,
                 AppointmentId = x.AppointmentId,
                 BookedByStaff = x.BookedByStaff,
                 CreatedAt = x.CreatedAt,
